+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 # 0.1.0
 
 ### Added
+
+- Add ETL to extract GTFS dataset